Added support for 16-bit and 32-bit read/writes to the RTC reg, and fixed the little-endian commands. With these fixes, the SPP menu background has now the right colour depending on the time (formerly it was always night blue).

+++ desmume/src/rtc.h
@@ -28,6 +28,6 @@
 #include "types.h"
 
 extern	void rtcInit();
-extern	u8 rtcRead();
+extern	u16 rtcRead();
 extern	void rtcWrite(u16 val);
 #endif
